+++ src/brasil/gov/newfieldcomplement/browser/static/newfieldcomplement.js
@@ -5,18 +5,4 @@ $(document).ready(function(){
         $('div#center-content-area').addClass('no-right-margin')
     }
 
-    //collapse
-    $('.filter-btn').click(function () {
-        var container = $('#collapse-form');
-
-        if (container.hasClass('show')) {
-            container.removeClass('show');
-            return false;
-        }
-
-        container.addClass('show');
-        return false;
-    });
-
-
 });